Factor out heat flux models for easier referencing
